drivers: npmx: shell: move timer code to the separate file
