+++ drivers/npmx/CMakeLists.txt
@@ -56,6 +56,7 @@ if(CONFIG_NPMX_SHELL)
     zephyr_library_sources(shell/led.c)
     zephyr_library_sources(shell/pof.c)
     zephyr_library_sources(shell/ship.c)
+    zephyr_library_sources(shell/timer.c)
 endif()
 
 # Get npmx include file
